+++ tests/__init__.py
@@ -8,4 +8,5 @@
 # TODO tplot psd
 # TODO dr surface area
 # TODO General Adsorbent Library and Evaluation (GALE)
-# TODO: does the langmuir area require relative pressure?
+# TODO does the langmuir area require relative pressure?
+# TODO better units for t-plot analysis, should follow isotherm units