+++ tests/characterisation/test_models_thickness.py
@@ -20,21 +20,13 @@ import pygaps.utilities.exceptions as pgEx
 @pytest.mark.characterisation
 class TestThicknessModels():
     """Test the thickness models."""
-    @pytest.mark.parametrize(
-        'model, thickness', [
-            (mt._THICKNESS_MODELS['Halsey'], [0.46, 0.62, 1.28]),
-            (mt._THICKNESS_MODELS['Harkins/Jura'], [0.37, 0.57, 1.32]),
-        ]
-    )
-    def test_static_models(self, model, thickness):
-        """Test each model against pre-calculated values."""
-        for index, value in enumerate([0.1, 0.4, 0.9]):
-            assert model(value) == pytest.approx(thickness[index], 0.01)
-
-    def test_get_thickness(self):
+    def test_get_thickness_model(self):
         """Get a regular model."""
-        assert mt.get_thickness_model('Halsey'
-                                      ) == mt._THICKNESS_MODELS['Halsey']
+        assert mt.get_thickness_model('Halsey') == mt._THICKNESS_MODELS['Halsey']
+
+    def test_get_thickness_model_precalc(self):
+        """Get a pre-calculated isotherm model."""
+        assert mt.get_thickness_model('SiO2 Jaroniec/Kruk/Olivier')
 
     def test_get_thickness_error(self):
         """When the model requested is not found we raise."""
@@ -48,3 +40,18 @@ class TestThicknessModels():
 
         ret = mt.get_thickness_model(call_this)
         assert ret() == 'called'
+
+    @pytest.mark.parametrize(
+        'modelname, thickness', [
+            ('Halsey', [0.46, 0.62, 1.28]),
+            ('Harkins/Jura', [0.37, 0.57, 1.32]),
+            ('zero thickness', [0, 0, 0]),
+            ('SiO2 Jaroniec/Kruk/Olivier', [0.362, 0.538, 1.036]),
+            ('carbon black Kruk/Jaroniec/Gadkaree', [0.374, 0.565, 1.257]),
+        ]
+    )
+    def test_models(self, modelname, thickness):
+        """Test each model against pre-calculated values."""
+        model = mt.get_thickness_model(modelname)
+        for index, value in enumerate([0.1, 0.4, 0.9]):
+            assert model(value) == pytest.approx(thickness[index], 0.01)